Add Storage.java to deal with tasks in hard disk
