+++ src/main/java/NUSYapBot.java
@@ -14,92 +14,16 @@ public class NUSYapBot {
 
     
 
-    public static ArrayList<Task> getTaskList() {
-        //save task from hard disk to an arrayList
-        ArrayList<Task> taskList = new ArrayList<>();
-        try {
-            //load the task list saved previously
-            File f = new File(STORAGE_PATH);
-
-            if (!f.createNewFile()) {
-                Scanner s = new Scanner(f);
-                while (s.hasNext()) {
-                    String taskLine = s.nextLine();
-                    //Format: Type | T/F | title | other-var
-                    String[] taskDetail = taskLine.split(" \\| ");
-                    
-                    if (taskDetail[0].equals("T")) {
-                        taskList.add(new ToDo(taskDetail[2], 
-                                taskDetail[1].equals("T")));
-                        
-                    } else if (taskDetail[0].equals("D")) {
-                        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy HHmm");
-                        LocalDateTime date = LocalDateTime.parse(taskDetail[3], formatter);
-                        
-                        taskList.add(new Deadline(taskDetail[2], 
-                                date, taskDetail[1].equals("T")));
-
-                    } else if (taskDetail[0].equals("E")) {
-                        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy HHmm");
-                        LocalDateTime start = LocalDateTime.parse(taskDetail[3], formatter);
-                        LocalDateTime end = LocalDateTime.parse(taskDetail[4], formatter);
-
-                        taskList.add(new Event(taskDetail[2], start, end, taskDetail[1].equals("T")));
-                    }
-                    
-                }
-            } 
-            
-        } catch (IOException e) {
-            System.out.println(e);
-            
-        }
-        return taskList;
-    }
-
-    public static void addTask(FileWriter fw, Task task) throws IOException {
-        //Format: Type | T/F | title | other-var
-        String line = task.getType() + " | "  
-                + (task.getIsCompleted() ? "T" : "F")
-                + " | " + task.getTitle();
-        if (task.getType() == 'D') {
-            Deadline taskD = (Deadline) task;
-            String ddl = taskD.getDeadline().format(
-                    DateTimeFormatter.ofPattern("dd-MM-yyyy HHmm"));
-            line += " | " + ddl;
-        } else if (task.getType() == 'E') {
-            Event taskE = (Event) task;
-            String start = taskE.getStartDate().format(
-                    DateTimeFormatter.ofPattern("dd-MM-yyyy HHmm"));
-            String end = taskE.getEndDate().format(
-                    DateTimeFormatter.ofPattern("dd-MM-yyyy HHmm"));
-                    
-            line += " | " + start + " | " + end;
-        } 
-        fw.write(line + "\n");
-    }
-
-    public static void addNewTask(Task task) throws IOException {
-        FileWriter fw = new FileWriter(STORAGE_PATH, true);
-        addTask(fw, task);
-        fw.close();
-    }
-
-    public static void rewriteMemory(ArrayList<Task> taskList) throws IOException {
-        FileWriter fw = new FileWriter(STORAGE_PATH, false); //false indicate overwrite mode
-        for (Task task: taskList) {
-            addTask(fw, task);
-        }
-        fw.close();
-    }
-
     public static void main(String[] args) {
         boolean flag = true;
-        ArrayList<Task> taskList = getTaskList();
-        int pointer = taskList.size();
+        try {
+            ArrayList<Task> taskList = Memory.getTaskList(STORAGE_PATH);
+        } catch (IOException e) {
+            System.out.println("Memory retrieval error. Please try again.")
+        }
         
         Ui.printWelcomeMessage(taskList);
-        
+
         Scanner input = new Scanner(System.in);
 
         while (flag) {
@@ -109,6 +33,7 @@ public class NUSYapBot {
                 System.out.println("_________________________________");
                 Ui.printTaskList(taskList);
                 System.out.println("_________________________________");
+
             } else if (answer.equals("bye")) {
                 flag = false;
                 Ui.printGoodbyeMessage();
@@ -128,10 +53,10 @@ public class NUSYapBot {
                 }
             } else if (answer.startsWith("delete")) {
                 try {
-                    CommandHandler.delete(taskList, answer, pointer);
-                    pointer--;
+                    CommandHandler.delete(taskList, answer);
+          
                     //rewrite the hard disk file
-                    rewriteMemory(taskList);
+                    Memory.rewriteMemory(taskList, STORAGE_PATH);
 
                 } catch (LackingInputException | InvalidTaskException | IOException e) {
                     System.out.println(e);
@@ -142,8 +67,7 @@ public class NUSYapBot {
                 try {
                     Task newTask = CommandHandler.createTask(answer);
                     taskList.add(newTask);
-                    pointer++;
-                    addNewTask(newTask);
+                    Memory.addNewTask(newTask, STORAGE_PATH);
                     Ui.printAddTaskMessage(taskList);
 
                 } catch (LackingInputException | UnrecognisedCommandException | DateFormatException e) {
